feat(demo): generate demo attendance records (AttendanceMonth)
